+++ CHANGELOG.md
@@ -142,3 +142,9 @@
 * Support different storage volumes for different types of workers.
 * Add support for workflow steps that are implemented as Jupyter Notebooks (\#79).
 * Include flowapp to run workflow templates using streamlit GUI.
+
+
+### 0.9.1 - 2021-06-22
+
+* Add option to resolve references to external files when reading Yaml files.
+* Support inclusion and import of files in instructions markdown text.